Updated readme and codec

+++ avro-codec/src/main/java/org/springframework/cloud/stream/schema/SchemaRegistryClient.java
@@ -9,8 +9,8 @@ import org.apache.avro.Schema;
  */
 public interface SchemaRegistryClient {
 
-	public Long register(Schema schema);
-	public Schema fetch(Long id);
+	public Integer register(Schema schema);
+	public Schema fetch(Integer id);
 	public List<Schema> find(Schema schema);
 	public List<Schema> find(String fqn);
